rename strat -> strat-pure. Create canon.
Create canonicalization strategy
